Got the tag filtering working finally. But now adding tags seems to have broken (sigh!). Will have another go tomorrow, I'm tired now.

+++ src/org/rgcrichton/episodic/TagListAdapter.java
@@ -76,9 +76,9 @@ public class TagListAdapter extends SimpleCursorAdapter {
 				}
 				else {
 					if (isChecked) {
-						tagList.addTagFilter(tagRowId);
+						tagList.addTagFilter((int)tagRowId);
 					} else {
-						tagList.removeTagFilter(tagRowId);
+						tagList.removeTagFilter((int)tagRowId);
 					}
 				}
 			}
